Updated model to merge it into scvi-tools
